pyproject.toml
[tool.poetry]
name = "codebase-analyzer"
version = "0.1.0"
description = "A tool for analyzing Python codebases and their dependencies"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
networkx = "^3.1"
matplotlib = "^3.7.1"
astroid = "^3.0.1"
python-dotenv = "^1.0.0"
langchain = "^0.1.0"
langchain-openai = "^0.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.1"
isort = "^5.13.2"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
multi_line_output = 3 